Bug 517593: Numerous NPEs when quitting Eclipse with active debug session
Background jobs are continuing to try to operate even though the platform
is shutting down. Simply check DebugPlugin.getDefault() != null to
ensure that the platform is running.

Note that the first thing that a lot of the debug code does (for example
when firing events) is DebugPlugin.getDefault().<something> causing
an NPE.

Change-Id: I12791be90331fdb27971a3c894d253a254931ad5
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>